fix: remove unnecessary code in component test: fix all warnings and passing all tests
