+++ src/pages/tests/AboutUs.spec.js
@@ -1,10 +1,22 @@
+// Configurations
+import { mount, config } from "@vue/test-utils";
+import { createTestingPinia } from "@pinia/testing";
+
+import NavBar from "@/components/NavBar.vue";
+import NavFooter from "@/components/NavFooter.vue";
+import router from "@/router/router.js";
+
+config.global.components["nav-bar"] = NavBar;
+config.global.components["nav-footer"] = NavFooter;
+config.global.plugins.push(router);
+
+// Test
 import {
     describe,
     it,
     expect,
     beforeAll
 } from "vitest";
-import { mount } from "@vue/test-utils";
 
 import AboutUs from "@/pages/AboutUs.vue";
 
@@ -13,7 +25,11 @@ describe("AboutUs Tests", () => {
     let wrapper;
 
     beforeAll(() => {
-        wrapper = mount(AboutUs);
+        wrapper = mount(AboutUs, {
+            global: {
+                plugins: [createTestingPinia()]
+            }
+        });
     });
 
     it("Should render AboutUs view", () => {
@@ -23,6 +39,6 @@ describe("AboutUs Tests", () => {
     it("Should load image from URL", function () {
         expect(wrapper.find("img").attributes("src")).not.toBe("");
         expect(wrapper.find("img").attributes("src")).not.toBe(null);
-        expect(wrapper.find("img").attributes("src")).toBe("https://avatars.githubusercontent.com/u/61888719?v=4");
+        expect(wrapper.findAll("img")[1].wrapperElement.attributes["0"].value).toBe("https://avatars.githubusercontent.com/u/61888719?v=4");
     });
 });
