+++ src/pages/tests/LessonsPage.spec.js
@@ -1,12 +1,23 @@
+// Configurations
+import { mount, config } from "@vue/test-utils";
 import { createTestingPinia } from "@pinia/testing";
+
+import NavBar from "@/components/NavBar.vue";
+import NavFooter from "@/components/NavFooter.vue";
+import router from "@/router/router.js";
+
+config.global.components["nav-bar"] = NavBar;
+config.global.components["nav-footer"] = NavFooter;
+config.global.plugins.push(router);
+
+// Test
 import {
     describe,
     it,
     expect,
-    beforeAll
+    beforeAll,
+    vi
 } from "vitest";
-import { fn } from "@vitest/spy";
-import { mount } from "@vue/test-utils";
 
 import LessonsPage from "@/pages/LessonsPage.vue";
 import { useDaysStore } from "@/store/lessons/daysStore.js";
@@ -21,7 +32,7 @@ describe("LessonsPage Tests", () => {
     beforeAll(() => {
         wrapper = mount(LessonsPage, {
             global: {
-                plugins: [createTestingPinia({ createSpy: fn })]
+                plugins: [createTestingPinia()]
             }
         });
 
@@ -48,10 +59,10 @@ describe("LessonsPage Tests", () => {
     });
 
     it("Should run changeRouteLessons when click on button Aprende", () => {
-        const spy = fn();
+        const spy = vi.fn();
 
         wrapper.vm.changeRouteLessons = spy;
-        wrapper.find("button").trigger("click");
+        wrapper.findAll("button")[1].trigger("click");
 
         expect(spy).toHaveBeenCalled();
     });
