+++ src/pages/tests/UserSettings.spec.js
@@ -1,12 +1,22 @@
+// Configurations
+import { mount, config } from "@vue/test-utils";
+import { createTestingPinia } from "@pinia/testing";
+
+import NavBar from "@/components/NavBar.vue";
+import NavFooter from "@/components/NavFooter.vue";
+import router from "@/router/router.js";
+
+config.global.components["nav-bar"] = NavBar;
+config.global.components["nav-footer"] = NavFooter;
+config.global.plugins.push(router);
+
+// Test
 import {
     describe,
     it,
     expect,
     beforeEach
 } from "vitest";
-import { fn } from "@vitest/spy";
-import { createTestingPinia } from "@pinia/testing";
-import { mount } from "@vue/test-utils";
 
 import UserSettings from "@/pages/UserSettings.vue";
 import { useUserStore } from "@/store/user/userStore.js";
@@ -28,7 +38,7 @@ describe("UserSettings", () => {
         };
         wrapper = mount(UserSettings, {
             global: {
-                plugins: [createTestingPinia({ createSpy: fn })]
+                plugins: [createTestingPinia()]
             }
         });
